Se agrega la logica de datos para los articulos, se modifican los screen y widgets para adaptarlo al uso desde la base de datos. Se modifica data_init y database_manager para actualizar la base de datos Se agrega una configuración para validar existecia. Pasos a seguir. Validar que se pueda incrementar o decrementar conservando los decimales. y que el input de cantidad sea double

+++ lib/screens/finalizar_venta.dart
@@ -89,7 +89,7 @@ class _FinalizarVentaScreenState extends State<FinalizarVentaScreen> {
     return widget.carrito.fold(0.0, (total, item) => total + item.subtotal);
   }
 
-  int _calcularTotalArticulos() {
+  double _calcularTotalArticulos() {
     return widget.carrito.fold(0, (total, item) => total + item.cantidad);
   }
